Fixed variables format in Dockerfiles and added variables linter to the script.

+++ scripts/drevops/tests/lint.dockerfiles.sh
@@ -9,6 +9,8 @@ set -eu
 
 CUR_DIR="$(dirname "$(dirname "$(cd "$(dirname "${BASH_SOURCE[0]}")/.." && pwd)")")"
 
+[ ! -f "${CUR_DIR}/scripts/drevops/tests/vendor/bin/shell-var-lint" ] && composer --working-dir="${CUR_DIR}/scripts/drevops/tests" install
+
 targets=()
 while IFS= read -r -d $'\0'; do
   targets+=("${REPLY}")
@@ -20,10 +22,15 @@ done < <(
     -not -path "*vendor*" -not -path "*node_modules*" \
     -print0
 )
-targets+=("${CUR_DIR}/install")
 
 echo "==> Linting DrevOps scripts and tests in ${CUR_DIR}."
 for file in "${targets[@]}"; do
+  # Temp script until shfmt implement the support for formatting variables.
+  # @see https://github.com/mvdan/sh/issues/1029
+  if ! "${CUR_DIR}/scripts/drevops/tests/vendor/bin/shell-var-lint" "${file}"; then
+    exit 1
+  fi
+
   if [ -f "${file}" ]; then
     echo "Checking file ${file}"
     docker run --rm -i hadolint/hadolint <"${file}"
